+++ src/compose.ts
@@ -751,7 +751,11 @@ export const composeHandler = ({
 		(isHandleFn && hasDefaultHeaders) ||
 		maybeStream
 
+	let _afterResponse: string | undefined
+
 	const afterResponse = () => {
+		if (_afterResponse !== undefined) return _afterResponse
+
 		if (!hooks.afterResponse?.length && !hasTrace) return ''
 
 		let afterResponse = ''
@@ -783,7 +787,7 @@ export const composeHandler = ({
 
 		afterResponse += '})\n'
 
-		return afterResponse
+		return (_afterResponse = afterResponse)
 	}
 
 	const mapResponse = (r = 'r') => {
@@ -1623,6 +1627,8 @@ export const composeHandler = ({
 									hook.fn.name
 								)
 
+								fnLiteral += `c.response=c.responseValue=be\n`
+
 								if (!returning) {
 									fnLiteral += isAsync(hook.fn)
 										? `await e.afterHandle[${i}](c, be)\n`
@@ -1669,6 +1675,8 @@ export const composeHandler = ({
 					}
 
 					mapResponseReporter.resolve()
+
+					fnLiteral += afterResponse()
 
 					fnLiteral += encodeCookie()
 					fnLiteral += `return mapEarlyResponse(${saveResponse}be,c.set${mapResponseContext